Move iptables test to integration tests
